NIFI-9032 Refactoring HDFS processors in order to increase flexibility
